+++ request_test.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"bytes"
 	"encoding/json"
+	"errors"
 	"net/http"
 	"os"
 	"testing"
@@ -238,6 +239,7 @@ func TestPostFiles(t *testing.T) {
 	_, _ = w.Write(f)
 	w.Flush()
 	f2, _ := os.Open("test.txt")
+	defer f2.Close()
 	a.Data = map[string]string{
 		"key": "value",
 		"a":   "123",
@@ -278,4 +280,115 @@ func TestGzip(t *testing.T) {
 	assert.Equal(t, t2 != "", true)
 	assert.Equal(t, c2 != nil, true)
 	assert.Equal(t, d.Get("gzipped").MustBool(), true)
+}
+
+func currentIP(u string) (ip string) {
+	c := &http.Client{}
+	a := NewArgs(c)
+	a.Proxy = u
+	url := "http://httpbin.org/get"
+	resp, _ := Get(url, a)
+	d, _ := resp.Json()
+	defer resp.Body.Close()
+
+	return d.Get("origin").MustString()
+}
+func currentIPHTTPS(u string) (ip string) {
+	c := &http.Client{}
+	a := NewArgs(c)
+	a.Proxy = u
+	url := "https://httpbin.org/get"
+	resp, _ := Get(url, a)
+	d, _ := resp.Json()
+	defer resp.Body.Close()
+
+	return d.Get("origin").MustString()
+}
+
+func TestProxy(t *testing.T) {
+	ip := currentIP("")
+	httpProxyURL := os.Getenv("http_proxy_url")
+	httpsProxyURL := os.Getenv("https_proxy_url")
+	socks5ProxyURL := os.Getenv("socks5_proxy_url")
+
+	assert.Equal(t, currentIP(httpProxyURL) != ip, true)
+	assert.Equal(t, currentIP(httpsProxyURL) != ip, true)
+	// assert.Equal(t, currentIPHTTPS(httpsProxyURL) != ip, true)
+	assert.Equal(t, currentIP(socks5ProxyURL) != ip, true)
+}
+
+func TestBasicAuth(t *testing.T) {
+	c := &http.Client{}
+	a := NewArgs(c)
+	a.BasicAuth = BasicAuth{"user", "passwd"}
+	url := "http://httpbin.org/basic-auth/user/passwd"
+	resp, _ := Get(url, a)
+	assert.Equal(t, resp.OK(), true)
+
+	a.BasicAuth = BasicAuth{
+		Username: "user2",
+		Password: "passwd2",
+	}
+	url = "http://httpbin.org/basic-auth/user2/passwd2"
+	resp, _ = Get(url, a)
+	assert.Equal(t, resp.OK(), true)
+}
+
+func TestResponseURL(t *testing.T) {
+	c := &http.Client{}
+	a := NewArgs(c)
+	url := "http://httpbin.org/get"
+	resp, _ := Get(url, a)
+	u, _ := resp.URL()
+	assert.Equal(t, u.String(), url)
+
+	url = "http://httpbin.org/redirect/3"
+	resp, _ = Get(url, a)
+	u, _ = resp.URL()
+	assert.Equal(t, u.String(), "http://httpbin.org/get")
+	url = "http://httpbin.org/redirect/3"
+
+	c.CheckRedirect = func(req *http.Request, via []*http.Request) error {
+		return errors.New("redirect")
+	}
+	resp, _ = Get(url, a)
+	u, _ = resp.URL()
+	assert.Equal(t, u.String(), "http://httpbin.org/redirect/2")
+
+}
+
+func TestCheckRedirect(t *testing.T) {
+	c := &http.Client{}
+	a := NewArgs(c)
+	url := "http://httpbin.org/get"
+	resp, _ := Get(url, a)
+	u, _ := resp.URL()
+	assert.Equal(t, u.String(), url)
+
+	url = "http://httpbin.org/redirect/3"
+	resp, _ = Get(url, a)
+	u, _ = resp.URL()
+	assert.Equal(t, u.String(), "http://httpbin.org/get")
+
+	url = "http://httpbin.org/redirect/15"
+	resp, _ = Get(url, a)
+	u, _ = resp.URL()
+	assert.Equal(t, u.String(), "http://httpbin.org/redirect/4")
+
+	url = "http://httpbin.org/redirect/2"
+	a.Headers = map[string]string{
+		"Referer": "http://example.com",
+	}
+	resp, _ = Get(url, a)
+	u, _ = resp.URL()
+	referer := resp.Request.Header.Get("Referer")
+	assert.Equal(t, u.String(), "http://httpbin.org/get")
+	assert.Equal(t, referer, "http://httpbin.org/redirect/1")
+	assert.Equal(t, resp.Request.Header.Get("User-Agent"), defaultUserAgent)
+
+	url = "http://httpbin.org/redirect/12"
+	DefaultRedirectLimit = 16
+	resp, _ = Get(url, a)
+	u, _ = resp.URL()
+	assert.Equal(t, u.String(), "http://httpbin.org/get")
 }